requirements.txt
fastapi==0.116.1
grpcio==1.74.0
grpcio-tools==1.74.0
uvicorn==0.35.0
